pyproject.toml
[project]
name = "google-ai-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.38.0",
    "python-dotenv>=1.1.1",
]
